+++ src/index.ts
@@ -33,6 +33,7 @@ export { default as Badge } from '@/package/components/badge.vue';
 export { default as Avatar } from '@/package/components/avatar.vue';
 export { default as DataTable } from '@/package/components/data-table.vue';
 export { default as Slider } from '@/package/components/slider.vue';
+export { default as Breadcrumbs } from '@/package/components/breadcrumbs.vue';
 export type { DataTableColumn } from '@/package/components/data-table.vue';
 // config
 export { default as VueForge, setTheme, updateTheme, getTheme } from '@/package/config/index';